% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aat_alpha.R
\name{aat_alpha}
\alias{aat_alpha}
\alias{aat_alpha_jackknife}
\alias{print.aat_alpha}
\alias{plot.aat_alpha}
\alias{print.aat_alpha_jackknife}
\alias{plot.aat_alpha_jackknife}
\title{Compute reliability using Cronbach's Alpha}
\usage{
aat_alpha(
  ds,
  subjvar,
  stimvar,
  pullvar,
  rtvar,
  algorithm = c("aat_singlemeandiff", "aat_singlemediandiff"),
  delete.missing = c("estimate", "subjects", "stimuli", "both", "none"),
  iters = 100
)

aat_alpha_jackknife(
  ds,
  subjvar,
  stimvar,
  pullvar,
  rtvar,
  algorithm = c("aat_singlemeandiff", "aat_singlemediandiff"),
  delete.missing = c("estimate", "subjects", "stimuli", "both", "none")
)

\method{print}{aat_alpha}(x, ...)

\method{plot}{aat_alpha}(x, ...)

\method{print}{aat_alpha_jackknife}(x, ...)

\method{plot}{aat_alpha_jackknife}(x, exclusions = c("both", "stimulus", "participant"), ...)
}
\arguments{
\item{ds}{a data.frame}

\item{subjvar}{character naming the column with participant IDs}

\item{stimvar}{character naming the column with stimulus IDs}

\item{pullvar}{character naming the column for movement direction}

\item{rtvar}{character indicating the column with reaction times}

\item{algorithm}{character name of the algorithm used to compute
per-stimulus approach bias scores}

\item{delete.missing}{character denoting the strategy to be taken when missing
values are encountered, which may occur when one participant saw a stimulus which
another participant did not see, or when all approach trials for a certain stimulus
were excluded for one participant. Defaults to estimating the covariance matrix.}

\item{iters}{Numeric. When there are missing values, the covariance matrix is estimated
instead of computed; how many iterations should be used?}

\item{x}{an \code{aat_alpha} or \code{aat_alpha_jackknife} object}

\item{...}{Ignored.}

\item{exclusions}{Character. Should the function display Cronbach's alpha for
individually excluded participants, stimuli, or both?}
}
\value{
\code{aat_alpha} returns an \code{aat_alpha} S3 object,
\code{aat_alpha_jackknife} returns an \code{aat_alpha_jackknife} S3 object,
both with print and plot methods.
}
\description{
\code{aat_alpha} computes approach-avoidance scores for each stimulus
within each participant. It then computes Cronbach's alpha by treating each stimulus
as an item on a questionnaire and measuring how much these scores correlate within
each participant.

\code{aat_alpha_jackknife} additionally leaves out one stimulus or participant
at a time and computes Cronbach's alpha for each such exclusion. This gives a
glimpse of which stimuli may have especially confused participants or otherwise
induced responses that are unlike those to other stimuli.
}
\details{
When there are missing values, the covariance matrix is estimated using
100 iterations of an expectation-maximization algorithm developed by Beale and Little (1975)
and implemented in R by StackExchange user Eric at URL
https://stats.stackexchange.com/questions/182718/ml-covariance-estimation-from-expectation-maximization-with-missing-data
}
\examples{
data(erotica)
#We artificially reduce the number of stimuli here because the original
#erotica dataset is not suitable for computing Cronbach's alpha.
erotica$stimulus<- substr(as.character(erotica$stimulus),5,5)

myalpha<-aat_alpha(erotica,"subject","stimulus","is_pull","RT")
print(myalpha)
plot(myalpha)

myalpha2<-aat_alpha_jackknife(erotica,"subject","stimulus","is_pull","RT")
print(myalpha2)
plot(myalpha2)
}
\references{
Cousijn, J., Goudriaan, A. E., & Wiers, R. W. (2011).
Reaching out towards cannabis: Approach‐bias in heavy cannabis users predicts
changes in cannabis use. Addiction, 106(9), 1667-1674.

Beale, E. M. L., & Little, R. J. A.. (1975). Missing Values in Multivariate Analysis.
Journal of the Royal Statistical Society. Series B (methodological), 37(1), 129–145.
}
\author{
Sercan Kahveci
}
