% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AATtools.R
\name{aat_compute}
\alias{aat_compute}
\title{Compute simple AAT scores}
\usage{
aat_compute(ds, subjvar, pullvar, targetvar, rtvar,
  algorithm = c("aat_doublemeandiff", "aat_doublemediandiff",
  "aat_dscore", "aat_dscore_multiblock", "aat_multilevelscore"),
  trialdropfunc = c("prune_nothing", "trial_prune_3SD",
  "trial_prune_SD_dropcases"), errortrialfunc = c("prune_nothing",
  "error_replace_blockmeanplus"), ...)
}
\arguments{
\item{ds}{a long-format data.frame}

\item{subjvar}{column name of subject variable}

\item{pullvar}{column name of pull/push indicator variable, must be numeric or logical (where pull is 1 or TRUE)}

\item{targetvar}{column name of target stimulus indicator, must be numeric or logical (where target is 1 or TRUE)}

\item{rtvar}{column name of reaction time variable}

\item{algorithm}{name of preferred computation algorithm, see \link{aat_doublemeandiff}}

\item{trialdropfunc}{preferred trial exclusion method (if any)}

\item{errortrialfunc}{preferred error recoding method (if any)}

\item{...}{}
}
\description{
Compute simple AAT scores, with optional outlier exclusion and error trial recoding.
}
