% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercat-alpha.R
\name{aat_intercategorical_alpha}
\alias{aat_intercategorical_alpha}
\title{Compute the reliability of a double difference score}
\usage{
aat_intercategorical_alpha(
  ds,
  subjvar,
  stimvar,
  pullvar,
  targetvar,
  rtvar,
  iters = 5
)
}
\arguments{
\item{ds}{the \code{data.frame} to use}

\item{subjvar}{Name of the subject-identifying variable}

\item{stimvar}{Name of the stimulus-identifying variable}

\item{pullvar}{Name of the movement-direction identifying variable}

\item{targetvar}{Name of the stimulus-category identifying variable}

\item{rtvar}{Name of the reaction-time identifying variable}

\item{iters}{If there are missing values (which is almost inevitable) then
multiple imputation will be used to complete the covariance matrix - this option sets
the number of multiple imputations to be used.}
}
\value{
Returns an \code{aat_intercategorical_alpha} object containing the alpha value
as well as the dataset and covariance matrix with replaced missing values.
}
\description{
This function uses Lord's (1963) algorithm to quickly compute the reliability
of a double mean difference score. Missing values are dealt with using multiple imputation.
}
\details{
When the study in question uses double median difference scores or D-scores,
\code{aat_splithalf()} is recommended instead; and if only single mean difference scores
are used, \code{aat_alpha} suffices.
}
\examples{
#We generate a dataset with 16 stimuli in each category
ds<-aat_simulate(biasfx_jitter=40,nstims=16)
ds$stim<-paste0(ds$stim,"-",ds$is_target)

#If the intercategorical alpha and
# bootstrapped splithalf measure something similar,
# then the outcomes should be close to each other.
aat_intercategorical_alpha(ds=ds,subjvar="subj",stimvar="stim",pullvar="is_pull",
                           targetvar="is_target",rtvar="rt")
aat_splithalf(ds=ds,subjvar="subj",pullvar="is_pull",targetvar="is_target",rtvar="rt",
              algorithm="aat_doublemeandiff",iters=100,plot=FALSE)
}
\references{
Lord, F.Y. (1963), "Elementary Models for Measuring Change",
in Problems in Measuring Change, C.W. Harris, ed.. Madison. Wisconsin:
University of Wisconsin.
}
