% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cortools.R
\name{cormean}
\alias{cormean}
\alias{cormean2}
\title{Compute a minimally biased average of correlation values}
\usage{
cormean(r, n, na.rm = F)

cormean2(r)
}
\arguments{
\item{r}{a vector containing correlation values}

\item{n}{a single value or vector containing sample sizes}

\item{na.rm}{Logical. Should missing values be removed?}
}
\value{
An average correlation.
}
\description{
This function computes a minimally biased average of correlation values.
This is needed because simple averaging of correlations is negatively biased,
and the often used z-transformation method of averaging correlations is positively biased.
The algorithm was developed by Olkin & Pratt (1958) and implemented by Jan Seifert.
}
\examples{
cormean(c(0,.3,.5),c(30,30,60))
}
\references{
Olkin, I., & Pratt, J. (1958). Unbiased estimation of certain correlation coefficients.
The Annals of Mathematical Statistics, 29. https://doi.org/10.1214/aoms/1177706717

https://github.com/SigurdJanson/AveragingCorrelations/blob/master/CorrAggBias.R

https://medium.com/@jan.seifert/averaging-correlations-part-ii-9143b546860b
}
