% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aat_alpha.R
\name{covmat-tools}
\alias{covmat-tools}
\alias{calpha}
\alias{covEM}
\title{Covariance Matrix Tools
These functions allow for the computation of covariance matrices with missing data
and for the computation of Cronbach's alpha from them.}
\usage{
calpha(covmat)

covEM(dat_missing, iters = 1000)
}
\arguments{
\item{covmat}{a covariance matrix}

\item{dat_missing}{a matrix with missing values}

\item{iters}{the number of iterations to perform to estimate missing values}
}
\description{
Covariance Matrix Tools
These functions allow for the computation of covariance matrices with missing data
and for the computation of Cronbach's alpha from them.
}
\section{Functions}{
\itemize{
\item \code{calpha}: Compute Cronbach's alpha from a covariance matrix

\item \code{covEM}: This function computes a covariance matrix from data with some values missing at random.
The code was written by Eric from StackExchange. https://stats.stackexchange.com/questions/182718/ml-covariance-estimation-from-expectation-maximization-with-missing-data
}}

\examples{
# make data with missing values
missing_mtcars <- mtcars
for(i in 1:20){
  missing_mtcars[sample(1:nrow(mtcars),1),sample(1:ncol(mtcars),1)]<-NA
}
covmat<-covEM(as.matrix(missing_mtcars))$sigma
calpha(covmat)
}
\references{
Beale, E. M. L., & Little, R. J. A.. (1975). Missing Values in Multivariate Analysis. Journal of the Royal Statistical Society. Series B (methodological), 37(1), 129–145.
}
