% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AATtools.R
\name{plot.aat_splithalf}
\alias{plot.aat_splithalf}
\title{Plot split-half scatterplots}
\usage{
\method{plot}{aat_splithalf}(x, type = c("median", "minimum", "maximum", "random"))
}
\arguments{
\item{x}{an \code{aat_splithalf} object}

\item{type}{Character argument indicating which iteration should be chosen. Must be an abbreviation of
\code{"median"} (default), \code{"minimum"}, \code{"maximum"}, or \code{"random"}.}
}
\description{
Plot split-half scatterplots
}
\examples{
#Coming soon
}
