% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relcorrections.R
\name{reliability-corrections}
\alias{reliability-corrections}
\alias{SpearmanBrown}
\alias{FlanaganRulon}
\title{Correct a correlation coefficient for being based on only a subset of the data}
\usage{
SpearmanBrown(
  corr,
  ntests = 2,
  fix.negative = c("nullify", "bilateral", "none")
)

FlanaganRulon(
  x1,
  x2,
  x = NULL,
  fix.negative = c("nullify", "bilateral", "none")
)
}
\arguments{
\item{corr}{To-be-corrected correlation coefficient}

\item{ntests}{An integer indicating how many times larger the full test is, for which the corrected correlation coefficient is being computed.
When \code{ntests=2}, the formula will compute what the correlation coefficient would be if the test were twice as long.}

\item{fix.negative}{Determines how to deal with a negative value. "nullify" sets it to zero,
"bilateral" applies the correction as if it were a positive number, and then sets it to negative.
"none" gives the raw value. It should be noted that negative values are not supposed to occur,
and there is no commonly accepted way to deal with them when they do occur.}

\item{x1}{scores from half 1}

\item{x2}{scores from half 2}

\item{x}{scores from the full sample (more accurate if provided)}
}
\value{
Spearman-Brown-corrected correlation coefficient.
}
\description{
Correct a correlation coefficient for being based on only a subset of the data
}
\section{Functions}{
\itemize{
\item \code{SpearmanBrown}: Perform a Spearman-Brown correction on the provided correlation score.

\item \code{FlanaganRulon}: Compute the true reliability using the Flanagan-Rulon formula,
which takes into account inequal variances between split halves
}}

\examples{

SpearmanBrown(.5)
FlanaganRulon(a<-rnorm(50),rnorm(50)+a*.5,fix.negative="bilateral")
}
